+++ src/Histogram.cpp
@@ -221,7 +221,7 @@ double N2D2::Histogram::calibrateKLDivergence(std::size_t nbBits) const {
     }
 
     const bool isUnsigned = mMinVal >= 0.0;
-    const std::size_t nbQuantizedBins = 1 << nbBits;
+    const std::size_t nbQuantizedBins = static_cast<std::size_t>(1) << nbBits;
 
     double threshold = Utils::max_abs(mMinVal, mMaxVal);
     double bestThreshold = threshold;
